+++ DataRepo/tests/test_exceptions.py
@@ -4,6 +4,7 @@ from DataRepo.utils.exceptions import (
     AggregatedErrors,
     MultiLoadStatus,
     UnexpectedIsotopes,
+    UnitsWrong,
     summarize_int_list,
 )
 
@@ -218,6 +219,27 @@ class ExceptionTests(TracebaseTestCase):
         esl = ["1-3", "5", "7-8"]
         sl = summarize_int_list(il)
         self.assertEqual(esl, sl)
+
+    def test_units_wrong(self):
+        units_dict = {
+            "Infusion Rate": {
+                "example_val": "3.3 non/sense",
+                "expected": "ul/m/g",
+                "rows": [5, 6],
+                "units": "non/sense",
+            },
+        }
+        uw = UnitsWrong(units_dict)
+        self.assertEqual(
+            (
+                "Unexpected units were found in 1 columns:\n"
+                "\tInfusion Rate (example: [3.3 non/sense] does not match units: [ul/m/g] on row(s): [5, 6])\n"
+                "Units are not allowed, but these also appear to be the wrong units."
+            ),
+            str(uw),
+        )
+        self.assertTrue(hasattr(uw, "units_dict"))
+        self.assertEqual(uw.units_dict, units_dict)
 
 
 class MultiLoadStatusTests(TracebaseTestCase):